feat: add fix testscript api to fix users testcases
